# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CurvePointResponse": {
        "properties": {
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "CurvePointResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IVSurfaceResponse": {
        "properties": {
          "maturities": {
            "items": {
              "type": "number"
            },
            "title": "Maturities",
            "type": "array"
          },
          "strikes": {
            "items": {
              "type": "number"
            },
            "title": "Strikes",
            "type": "array"
          },
          "surface": {
            "title": "Surface",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          },
          "values": {
            "items": {
              "items": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "array"
            },
            "title": "Values",
            "type": "array"
          }
        },
        "required": [
          "ticker",
          "surface",
          "strikes",
          "maturities",
          "values"
        ],
        "title": "IVSurfaceResponse",
        "type": "object"
      },
...
      "MarketHistoryPoint": {
        "properties": {
          "close": {
            "title": "Close",
            "type": "number"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          }
        },
        "required": [
          "date",
          "close"
        ],
        "title": "MarketHistoryPoint",
        "type": "object"
      },
      "MarketHistoryResponse": {
        "properties": {
          "points": {
            "items": {
              "$ref": "#/components/schemas/MarketHistoryPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "points"
        ],
        "title": "MarketHistoryResponse",
        "type": "object"
      },
...
      "RatesCurveResponse": {
        "properties": {
          "curve": {
            "title": "Curve",
            "type": "string"
          },
          "zero": {
            "items": {
              "$ref": "#/components/schemas/CurvePointResponse"
            },
            "title": "Zero",
            "type": "array"
          }
        },
        "required": [
          "curve",
          "zero"
        ],
        "title": "RatesCurveResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "quantModeling API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/market/iv/surface": {
      "get": {
        "operationId": "iv_surface",
        "parameters": [
          {
            "in": "query",
            "name": "ticker",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "surface",
            "required": false,
            "schema": {
              "default": "mid",
              "pattern": "^(mid|bid|ask)$",
              "title": "Surface",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IVSurfaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Iv Surface"
      }
    },
    "/market/prices/history": {
      "get": {
        "operationId": "market_history",
        "parameters": [
          {
            "in": "query",
            "name": "ticker",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Ticker",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "range",
            "required": false,
            "schema": {
              "default": "6M",
              "pattern": "^(1M|3M|6M|YTD|1Y|2Y)$",
              "title": "Range",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Market History"
      }
    },
    "/market/rates/curve": {
      "get": {
        "operationId": "rates_curve",
        "parameters": [
          {
            "in": "query",
            "name": "curve",
            "required": false,
            "schema": {
              "default": "Treasury",
              "pattern": "^(Treasury|SOFR|FedFunds)$",
              "title": "Curve",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "curve_type",
            "required": false,
            "schema": {
              "default": "zero",
              "pattern": "^(zero|forward)$",
              "title": "Curve Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "fixed_period_years",
            "required": false,
            "schema": {
              "default": 0.5,
              "exclusiveMinimum": 0.0,
              "maximum": 10.0,
              "title": "Fixed Period Years",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatesCurveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rates Curve"
      }
    },
...
  }
}